Added ratings

+++ manager.py
@@ -3,6 +3,8 @@ import time
 import atexit, signal
 import argparse
 import shutil
+from gooey import Gooey
+import json
 
 
 def clear_temp_files(shared_drive_path):
@@ -12,14 +14,22 @@ def clear_temp_files(shared_drive_path):
 
     # Clear previous files
     for folder in ["source", "dicom", "nifti", "aligned"]:
-        for root, dirs, files in os.walk(f"{shared_drive_path}/data/{folder}"):
-            for f in files:
-                os.unlink(os.path.join(root, f))
-            for d in dirs:
-                shutil.rmtree(os.path.join(root, d))
+
+        path = f"{shared_drive_path}/data/{folder}"
+
+        # Create folders if not present
+        if not os.path.exists(path):
+            os.makedirs(path)
+        else:
+            for root, dirs, files in os.walk(path):
+                for f in files:
+                    os.unlink(os.path.join(root, f))
+                for d in dirs:
+                    shutil.rmtree(os.path.join(root, d))
 
 
-if __name__ == "__main__":
+@Gooey
+def main():
 
     '''
     This script is a parent for the processor and collector subprocesses.
@@ -30,50 +40,49 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='')
 
     parser.add_argument('participant', type=int, help="Participant number.")
-    parser.add_argument('session', type=int, help="Real-time session number.")
+    parser.add_argument('session', type=int, help="Session number.")
     parser.add_argument('run', type=int, help="Run index for this session.")
-    parser.add_argument('target', type=str, help="Evolution target. Example options:\n'fear'\n'cute'\n'disgust'\n'neurosynth'")
-    parser.add_argument('--simulate', action='store_true', help="Is this run being simulated offline?", required=False)
-    parser.add_argument('--shared_drive_path', type=str, help="Alternative shared drive path.", required=False)
-    parser.add_argument('--diffusion_steps', type=int, default=25, help="Number of diffusion steps per image", required=False)
+    parser.add_argument('condition', type=str, choices=['brain', 'ratings'], help="Are brain scores or self-report ratings being used?")
 
+    parser.add_argument('--target', type=str, default="fear", help="Name of target map (e.g., input 'fear' if file is 'fear_map.nii')")
+    parser.add_argument('--simulate', action='store_true', help="Is this run being simulated offline?", required=False)
     
     args = parser.parse_args()
 
     participant = args.participant
     ses = args.session
     run = args.run
+    condition = args.condition
     target = args.target
-    diffusion_steps = args.diffusion_steps
 
-    # Set paths
-    if args.shared_drive_path:
-        shared_drive_path = args.shared_drive_path
-    else:
-        shared_drive_path = r"C:\Users\TD_Lab\Desktop\Realtime\shared_drive"
+    with open('config.json') as f:
+        config = json.load(f)
+
+    shared_drive_path = config["shared_drive_path"]
 
     output_path = f"{shared_drive_path}/images/sub-{participant:02}/ses-{ses:02}/run-{run}"
-
 
     # Clear existing files, if any
     clear_temp_files(shared_drive_path)
 
 
-    # Simulator
-    if args.simulate:
-        sim = subprocess.Popen(["python", "siemens_simulator.py", shared_drive_path])
-        atexit.register(os.kill, sim.pid, signal.CTRL_C_EVENT)
-        time.sleep(2)
+    if condition == "brain":
+        # Simulation
+        if args.simulate:
+            sim = subprocess.Popen(["python", "siemens_simulator.py", shared_drive_path])
+            atexit.register(os.kill, sim.pid, signal.CTRL_C_EVENT)
+            time.sleep(2)
 
+        if condition == "brain":
+            # Start processor subprocess
+            proc = subprocess.Popen(["python", "processor.py", shared_drive_path, str(participant)])
+            atexit.register(os.kill, proc.pid, signal.CTRL_C_EVENT)
 
-    # Start processor subprocess
-    proc = subprocess.Popen(["python", "processor.py", shared_drive_path, str(participant)])
-    atexit.register(os.kill, proc.pid, signal.CTRL_C_EVENT)
-    
 
     # Start collector subprocess
-    collector = subprocess.Popen(['python', 'collector.py', shared_drive_path, output_path, str(participant), target, str(diffusion_steps)])#, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    collector = subprocess.Popen(['python', 'collector.py', shared_drive_path, output_path, str(participant), target, condition])#, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     atexit.register(os.kill, collector.pid, signal.CTRL_C_EVENT)
+
 
     while(True):
     
@@ -85,6 +94,12 @@ if __name__ == "__main__":
             # Clear existing files, if any
             clear_temp_files(shared_drive_path)
             exit()
+
+    return None
+
+if __name__ == "__main__":
+    main()
+
